+++ engine/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.17.0-dev
 * Cleanup RTC Engine deps. Move metrics to the WebRTC Endpoint [#306](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/306)
+* Add `get_num_forwarded_tracks` function [#300](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/300/)
+
 
 ## 0.16.0
 * Convert RTC Engine into monorepo [#298](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/298)