+++ hls/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.2.0-dev
+* Rename FileEndpoint to FileSourceEndpoint [#300](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/300/)
+* Move StaticTrackSender to membrane_rtc_engine [#300](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/300/)
 
 ## 0.1.0
 * Initial release